backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
faiss-cpu==1.7.4
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
pytest==7.4.3

